+++ src/gdalraster.cpp
@@ -771,6 +771,154 @@ void GDALRaster::fillRaster(int band, double value, double ivalue) {
 	}
 }
 
+SEXP GDALRaster::getColorTable(int band) const {
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+	
+	GDALRasterBandH hBand = this->_getBand(band);
+	GDALColorTableH hColTbl = GDALGetRasterColorTable(hBand);
+	if (hColTbl == NULL)
+		return R_NilValue;
+
+	int nEntries = GDALGetColorEntryCount(hColTbl);
+	GDALPaletteInterp gpi = GDALGetPaletteInterpretation(hColTbl);
+	Rcpp::IntegerMatrix col_tbl(nEntries, 5);
+	Rcpp::CharacterVector col_tbl_names;
+
+	if (gpi == GPI_Gray) {
+		col_tbl_names = {"value", "gray", "c2", "c3", "c4"};
+		Rcpp::colnames(col_tbl) = col_tbl_names;
+	}
+	else if (gpi == GPI_RGB) {
+		col_tbl_names = {"value", "red", "green", "blue", "alpha"};
+		Rcpp::colnames(col_tbl) = col_tbl_names;
+	}
+	else if (gpi == GPI_CMYK) {
+		col_tbl_names = {"value", "cyan", "magenta", "yellow", "black"};
+		Rcpp::colnames(col_tbl) = col_tbl_names;
+	}
+	else if (gpi == GPI_HLS) {
+		col_tbl_names = {"value", "hue", "lightness", "saturation", "c4"};
+		Rcpp::colnames(col_tbl) = col_tbl_names;
+	}
+	else {
+		col_tbl_names = {"value", "c1", "c2", "c3", "c4"};
+		Rcpp::colnames(col_tbl) = col_tbl_names;
+	}
+
+	for (int i=0; i < nEntries; ++i) {
+		const GDALColorEntry* colEntry = GDALGetColorEntry(hColTbl, i);
+		col_tbl(i, 0) = i;
+		col_tbl(i, 1) = colEntry->c1;
+		col_tbl(i, 2) = colEntry->c2;
+		col_tbl(i, 3) = colEntry->c3;
+		col_tbl(i, 4) = colEntry->c4;
+	}
+	
+	return col_tbl;
+}
+
+std::string GDALRaster::getPaletteInterp(int band) const {
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+	
+	GDALRasterBandH hBand = this->_getBand(band);
+	GDALColorTableH hColTbl = GDALGetRasterColorTable(hBand);
+	if (hColTbl == NULL)
+		return "";
+
+	GDALPaletteInterp gpi = GDALGetPaletteInterpretation(hColTbl);
+
+	if (gpi == GPI_Gray) {
+		return "Gray";
+	}
+	else if (gpi == GPI_RGB) {
+		return "RGB";
+	}
+	else if (gpi == GPI_CMYK) {
+		return "CMYK";
+	}
+	else if (gpi == GPI_HLS) {
+		return "HLS";
+	}
+	else {
+		return "unknown";
+	}
+}
+
+bool GDALRaster::setColorTable(int band, Rcpp::RObject &col_tbl, 
+		std::string palette_interp) {
+		
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+		
+	if (GDALGetAccess(hDataset) == GA_ReadOnly)
+		Rcpp::stop("Dataset is read-only.");
+	
+	GDALRasterBandH hBand = this->_getBand(band);
+	
+	Rcpp::IntegerMatrix m_col_tbl;
+	if (Rcpp::is<Rcpp::DataFrame>(col_tbl)) {
+		m_col_tbl = _df_to_int_matrix(col_tbl);
+	}
+	else if (Rcpp::is<Rcpp::IntegerVector>(col_tbl)) {
+		if (Rf_isMatrix(col_tbl))
+			m_col_tbl = Rcpp::as<Rcpp::IntegerMatrix>(col_tbl);
+	}
+	else {
+		Rcpp::stop("col_tbl must be a data frame or matrix.");
+	}
+		
+	if (m_col_tbl.ncol() < 4 || m_col_tbl.ncol() > 5)
+		Rcpp::stop("col_tbl must have four or five columns.");
+	if (m_col_tbl.ncol() == 4) {
+		Rcpp::IntegerVector c4(m_col_tbl.nrow(), 255);
+		m_col_tbl = Rcpp::cbind(m_col_tbl, c4);
+	}
+	
+	GDALPaletteInterp gpi;
+	if (palette_interp ==  "Gray" || palette_interp == "gray")
+		gpi = GPI_Gray;
+	else if (palette_interp ==  "RGB")
+		gpi = GPI_RGB;
+	else if (palette_interp ==  "CMYK")
+		gpi = GPI_CMYK;
+	else if (palette_interp ==  "HLS")
+		gpi = GPI_HLS;
+	else
+		Rcpp::stop("Invalid palette_interp.");
+	
+	int max_value = Rcpp::max(m_col_tbl.column(0));
+	GDALColorTableH hColTbl = GDALCreateColorTable(gpi);
+	// initialize all entries
+	for (int i=0; i < max_value; ++i) {
+		const GDALColorEntry col = {0, 0, 0, 0};
+		GDALSetColorEntry(hColTbl, i, &col);
+	}
+	
+	// set entries from input table
+	for (int i=0; i < m_col_tbl.nrow(); ++i) {
+		if (m_col_tbl(i,0) >= 0) {
+			const GDALColorEntry col = {
+					static_cast<short>(m_col_tbl(i,1)),
+					static_cast<short>(m_col_tbl(i,2)),
+					static_cast<short>(m_col_tbl(i,3)),
+					static_cast<short>(m_col_tbl(i,4)) };
+			GDALSetColorEntry(hColTbl, m_col_tbl(i,0), &col);
+		}
+		else {
+			Rcpp::warning("Warning: skipped entry with negative value.");
+		}
+	}
+	
+	CPLErr err = GDALSetRasterColorTable(hBand, hColTbl);
+	GDALDestroyColorTable(hColTbl);
+	if (err == CE_None)
+		return true;
+	else
+		return false;
+}
+
 int GDALRaster::getChecksum(int band, int xoff, int yoff,
 		int xsize, int ysize) const {
 
@@ -786,7 +934,9 @@ void GDALRaster::close() {
 	hDataset = NULL;
 }
 
+// ********************************************************
 // class methods for internal use not exposed in R
+// ********************************************************
 
 GDALRasterBandH GDALRaster::_getBand(int band) const {
 	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
@@ -883,6 +1033,12 @@ RCPP_MODULE(mod_GDALRaster) {
     	"Write a region of raster data for a band.")
     .method("fillRaster", &GDALRaster::fillRaster, 
     	"Fill this band with a constant value.")
+    .const_method("getColorTable", &GDALRaster::getColorTable, 
+    	"Return the color table associated with this band.")
+    .const_method("getPaletteInterp", &GDALRaster::getPaletteInterp, 
+    	"Get the palette interpretation.")
+    .method("setColorTable", &GDALRaster::setColorTable, 
+    	"Set a color table for this band.")
     .const_method("getChecksum", &GDALRaster::getChecksum, 
     	"Compute checksum for raster region.")
     .method("close", &GDALRaster::close,